+++ portainer-monitor-addon/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.0.2
+- Add `Docker Version` sensor for each Endpoint
+- Changed udpate label of Portainer update sensor from `Update Available` to `Update`
+- Improved removal of MQTT entities on addon stop
+
 ## 1.0.1
 - Fixed value type error inside Endpoint API